wip - try to get remote debugging from vscode working
there is an issue with launching chrome in debugging mode if there are any
current running instances. Was able to get past that by closing down all chromes
and launching manually from the cli with debugging turned on. Attach worked at
that point but still couldn't set breakpoints in vscode due to either paths
or existence of source maps not being right.